wip: first steps to address nikhils comment
